[DataPipe] Refactor _hook_iterator for readability
[ghstack-poisoned]
